Conflicts:
	RobotSplit/RobotSplit/Game.cpp

+++ RobotSplit/RobotSplit/BrainMagnet.cpp
@@ -15,11 +15,13 @@ BrainMagnet::BrainMagnet(std::string spriteName):
 	ropeSwing=true;
 }
 
-void BrainMagnet::setSolid(bool b){
+void BrainMagnet::setSolid(bool b)
+{
 	mSolid=b;
 }
 
-void BrainMagnet::update(){
+void BrainMagnet::update()
+{
 	mActiveAnimation->update();
 
 	if(mRot>=45)
@@ -34,11 +36,11 @@ void BrainMagnet::update(){
 	{
 		if(mRot<20 && mRot>-20)
 		{
-			mRot+=1.7*0.5;
+			mRot+=(float)1.7*0.5;
 		}
 		else
 		{
-			mRot+=1.3*0.5;
+			mRot+=(float)1.3*0.5;
 		}
 		BrainMagnet::rotate(mRot);
 	}
@@ -46,25 +48,28 @@ void BrainMagnet::update(){
 	{
 		if(mRot>-20 && mRot<20)
 		{
-			mRot-=1.7*0.5;
+			mRot-=(float)1.7*0.5;
 		}
 		else
 		{
-			mRot-=1.3*0.5;
+			mRot-=(float)1.3*0.5;
 		}
 		BrainMagnet::rotate(mRot);
 	}
 
-	if(mBrainAni.getCurrentFrame()!=6){
+	if(mBrainAni.getCurrentFrame()!=6)
+	{
 		mActiveAnimation=&mBrainAni;
 		mSolid=false;
 	}
-	else{
+	else
+	{
 		mActiveAnimation=&mBrain;
 	}
 }
 
-sf::Sprite BrainMagnet::getSprite(){
+sf::Sprite BrainMagnet::getSprite()
+{
 	mActiveAnimation->setPosition(mPosition);
 	mSprite=mActiveAnimation->getSprite();
 	mSprite.setPosition(mSprite.getPosition()+sf::Vector2f(mSprite.getGlobalBounds().width/2, 10));
@@ -72,7 +77,8 @@ sf::Sprite BrainMagnet::getSprite(){
 	mSprite.setRotation(mRotation);
 	return mSprite;
 }
-void BrainMagnet::restartAnimations(){
+void BrainMagnet::restartAnimations()
+{
 	mBrainAni.restart();
 	mTimer.restart(); 
 	mActiveAnimation=&mBrainAni;
