+++ RobotSplit/RobotSplit/Laser.cpp
@@ -47,7 +47,7 @@ void Laser::update()
 sf::Sprite Laser::getSprite()
 {
 	mSprite=mAnimation->getSprite();
-	mSprite.setTextureRect(sf::IntRect(mSprite.getTextureRect().left,  mSprite.getTextureRect().top, mSprite.getTextureRect().width, mLength));
+	mSprite.setTextureRect(sf::IntRect(mSprite.getTextureRect().left,  mSprite.getTextureRect().top, mSprite.getTextureRect().width, (int)mLength));
 	mSprite.setPosition(mPosition);
 	//mSprite.setOrigin(-8, -4);
 	mSprite.setRotation(mRotation);
